+++ src/integration-helpers.ts
@@ -1,0 +1,25 @@
+import { createHash } from "node:crypto";
+import ts from "typescript";
+import { z } from "zod";
+
+export const f = ts.factory;
+export const exportModifier = [f.createModifier(ts.SyntaxKind.ExportKeyword)];
+
+export const ucFirst = (subject: string) =>
+  subject.charAt(0).toUpperCase() + subject.slice(1).toLowerCase();
+
+export const lcFirst = (subject: string) =>
+  subject.charAt(0).toLowerCase() + subject.slice(1);
+
+export const makeCleanId = (...args: string[]) =>
+  args
+    .flatMap((entry) => entry.split(/[^A-Z0-9]/gi)) // split by non-alphanumeric characters
+    .flatMap((entry) =>
+      // split by sequences of capitalized letters
+      entry.replaceAll(/[A-Z]+/g, (beginning) => `/${beginning}`).split("/"),
+    )
+    .map(ucFirst)
+    .join("");
+
+export const defaultSerializer = (schema: z.ZodTypeAny): string =>
+  createHash("sha1").update(JSON.stringify(schema), "utf8").digest("hex");
